requirements.txt
nicegui>=1.4.0
aiohttp>=3.8.0
tenacity>=8.0.0
python-dotenv>=0.19.0
typing-extensions>=4.0.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0